+++ app/assets/javascripts/mr_notes/stores/actions.js
@@ -1,7 +1,9 @@
 import types from './mutation_types';
 
-export default {
-  setActiveTab({ commit }, tab) {
-    commit(types.SET_ACTIVE_TAB, tab);
-  },
-};
+export function setActiveTab({ commit }, tab) {
+  commit(types.SET_ACTIVE_TAB, tab);
+}
+
+export function setEndpoints({ commit }, endpoints) {
+  commit(types.SET_ENDPOINTS, endpoints);
+}