+++ app/assets/javascripts/mr_notes/stores/modules/index.js
@@ -1,9 +1,10 @@
-import actions from '../actions';
+import * as actions from '../actions';
 import getters from '../getters';
 import mutations from '../mutations';
 
 export default () => ({
   state: {
+    endpoints: {},
     activeTab: null,
   },
   actions,