+++ app/assets/javascripts/mr_notes/stores/mutation_types.js
@@ -1,3 +1,4 @@
 export default {
   SET_ACTIVE_TAB: 'SET_ACTIVE_TAB',
+  SET_ENDPOINTS: 'SET_ENDPOINTS',
 };