+++ app/assets/javascripts/mr_notes/stores/mutations.js
@@ -4,4 +4,7 @@ export default {
   [types.SET_ACTIVE_TAB](state, tab) {
     Object.assign(state, { activeTab: tab });
   },
+  [types.SET_ENDPOINTS](state, endpoints) {
+    Object.assign(state, { endpoints });
+  },
 };