+++ spec/graphql/types/ci/job_type_spec.rb
@@ -8,6 +8,8 @@ RSpec.describe Types::Ci::JobType do
 
   it 'exposes the expected fields' do
     expected_fields = %i[
+      id
+      shortSha
       pipeline
       name
       needs